chore: move all card to aggregate all of them as UI cards
