+++ src/components/Cards/ProjectCard/components/ProjectDetailModal.types.ts
