+++ src/components/Cards/SkillCard/SkillCard.module.scss
