+++ src/components/Cards/SkillCard/SkillCard.types.ts
